# docs/web_deploy/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: nonce-auth-svc
  description: ''
servers:
  - url: 'https://api.demo.tokend.io'
    description: TokenD Developer Environment
paths:
...
  /integrations/nonce-auth-svc/validate:
    get:
      tags:
        - User Authentication
      summary: Validation
      operationId: validate
      description: Check if token is valid
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
            format: 'Bearer {token}'
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '204':
          description: Success
        '400':
          description: Bad Request
...